Renamed DOMError to FileError. Renamed ProgressEvent to FileProgressEvent and derived it from Event. Refactored. Removed BlobPropertyBag and changed the 3rd Blob constructor parameter to object. Changed FileList Item function to an indexer. Changed UShorts to Int. Created FileReadyState enum from those constants. FileInputElement is now sealed and has a private constructor. Removed the ScriptName metadata attribute. Created newlines at the end of all files. Moved FileAccess to Files and renamed the namespace.
