+++ src/Libraries/Web/Html/Window.cs
@@ -5,7 +5,7 @@
 
 using System;
 using System.Html.Data;
-using System.Html.Data.FileAccess;
+using System.Html.Data.Files;
 using System.Html.Data.IndexedDB;
 using System.Html.Data.Sql;
 using System.Runtime.CompilerServices;
